[BOOK-124] chore: code style check success
